requirements.txt
accelerate==1.8.1
altgraph==0.17.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.4
cssselect2==0.8.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.9.0
django-cors-headers==4.7.0
django-filter==25.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
future==0.18.2
h11==0.16.0
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
macholib==1.15.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
murmurhash==1.0.13
mysqlclient==2.2.7
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20250625
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0
sniffio==1.3.1
spacy==3.8.7
spacy-alignments==0.9.2
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.9
sqlparse==0.5.3
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.49.0
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.2
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
wrapt==1.17.2
zopfli==0.2.3.post1
